Fix new clippy lints for Rust 1.66
- https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
- https://rust-lang.github.io/rust-clippy/master/index.html#manual_clamp
- https://rust-lang.github.io/rust-clippy/master/#suboptimal_flops

+++ src/api/internal.rs
@@ -726,7 +726,7 @@ impl<T: Pixel> ContextInner<T> {
         output_frameno,
         segmentation: fs.segmentation,
       });
-      for i in 0..(REF_FRAMES as usize) {
+      for i in 0..REF_FRAMES {
         if (fi.refresh_frame_flags & (1 << i)) != 0 {
           coded_data.lookahead_rec_buffer.frames[i] = Some(Arc::clone(&rfs));
           coded_data.lookahead_rec_buffer.deblock[i] = fs.deblock;
@@ -808,7 +808,7 @@ impl<T: Pixel> ContextInner<T> {
       output_frameno,
       segmentation: fs.segmentation,
     });
-    for i in 0..(REF_FRAMES as usize) {
+    for i in 0..REF_FRAMES {
       if (fi.refresh_frame_flags & (1 << i)) != 0 {
         coded_data.lookahead_rec_buffer.frames[i] = Some(Arc::clone(&rfs));
         coded_data.lookahead_rec_buffer.deblock[i] = fs.deblock;
